Create new modules to separate dom from logic
